+++ include/Makefile.am
@@ -1,3 +1,9 @@
 ## $NetBSD: Makefile.am,v 1.2 2009/05/26 05:41:41 lukem Exp $
 
-include_HEADERS		= rnp.h
+include_HEADERS		= \
+    rnp.h \
+    rnp/rnp.h \
+    rnp/rnp_def.h \
+    rnp/rnp_obsolete_defs.h \
+    rnp/rnp_types.h \
+    rekey/rnp_key_store.h
