+++ include/rnp/rnp_def.h
