+++ include/rnp/rnp_obsolete_defs.h
