+++ src/lib/compress.c
@@ -75,7 +75,7 @@ __RCSID("$NetBSD: compress.c,v 1.23 2012/03/05 02:20:18 christos Exp $");
 #include <string.h>
 #include <stdlib.h>
 
-#include <rnp_def.h>
+#include <rnp/rnp_def.h>
 #include "packet-parse.h"
 #include "errors.h"
 #include "utils.h"
