+++ src/librekey/key_store_kbx.h
@@ -27,8 +27,8 @@
 #ifndef RNP_KEY_STORE_KBX_H
 #define RNP_KEY_STORE_KBX_H
 
-#include <rnp.h>
-#include <rnp_key_store.h>
+#include <rnp/rnp.h>
+#include <rekey/rnp_key_store.h>
 
 bool rnp_key_store_kbx_from_mem(pgp_io_t *, rnp_key_store_t *, pgp_memory_t *);
 bool rnp_key_store_kbx_to_mem(pgp_io_t *, rnp_key_store_t *, const uint8_t *, pgp_memory_t *);
