+++ src/tests/exportkey.c
@@ -25,8 +25,8 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <rnp.h>
-#include <rnp_key_store.h>
+#include <rnp/rnp.h>
+#include <rekey/rnp_key_store.h>
 #include "rnp_tests.h"
 #include "support.h"
