+++ src/tests/generatekey.c
@@ -24,8 +24,8 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <rnp.h>
-#include <rnp_key_store.h>
+#include <rnp/rnp.h>
+#include <rekey/rnp_key_store.h>
 #include "rnp_tests.h"
 #include "support.h"
 #include "symmetric.h"
